fix(deps): Update go module dependencies
This commit updates the go module dependencies for the `ramcache`, `memcache`, and `rediscluster` packages. It updates the `github.com/bartventer/gocache` dependency to version `v1.4.1` in all three packages. Additionally, it adds a temporary replacement for `github.com/docker/docker` to resolve an issue with the `testcontainers-go` package. Once the issue is resolved, the replacement should be removed.